fix: display no recommendation handled and remove debug logs
